Admin/Announcements endpoints
* List announcements
* Create announcement
* Delete announcement
* Update announcement

Extra:
* All Limit should be `uint` because they are all `0 < Limit <= 100`
* All Offset should be `uint64` because they are all `0 < Offset` and
  they can be large numbers.

Resolves #60

+++ services/notes/conversation.go
@@ -9,7 +9,7 @@ import (
 type ConversationRequest struct {
 	NoteID string `json:"noteId"`
 	Limit  uint   `json:"limit"`
-	Offset uint   `json:"offset,omitempty"`
+	Offset uint64 `json:"offset,omitempty"`
 }
 
 // Validate the request.
